POC for vehicle classification project
Classifying car/truck and some other image which is classified as unknown images. We couldn't find dataset for other vehicles. Once we will have the dataset for other vehicle, we can add it in easily.

Accuracy is 84 %,. We can get better accuracy if we tune the parameter. Due to lack of better system I didn't try for more accuracy, but this is not very difficult. We can easily achieve 97-98% of accuracy (which is equilvalent to human) by tuning the model if we have better system GPU/CPU